+++ src/App.tsx
@@ -1,13 +1,9 @@
 // import { useState } from "react";
-import { BrowserRouter as Router } from "react-router-dom";
-import RoutesComponent from "./components/Routes";
+import { RouterProvider } from "react-router";
+import Routes from "../src/components/Routes";
 
 function App() {
-  return (
-    <Router>
-      <RoutesComponent />
-    </Router>
-  );
+  return <RouterProvider router={Routes} />;
 }
 
 export default App;
