+++ src/components/Layout/Admin.tsx
@@ -34,7 +34,7 @@ function Admin() {
           className="dropdown-content menu bg-base-100 rounded-box z-1 w-52 p-2 mr-3 shadow-sm"
         >
           <li>
-            <a>
+            <Link to={"/dashboard"}>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 24 24"
@@ -45,10 +45,10 @@ function Admin() {
                 <path d="m12 5.432 8.159 8.159c.03.03.06.058.091.086v6.198c0 1.035-.84 1.875-1.875 1.875H15a.75.75 0 0 1-.75-.75v-4.5a.75.75 0 0 0-.75-.75h-3a.75.75 0 0 0-.75.75V21a.75.75 0 0 1-.75.75H5.625a1.875 1.875 0 0 1-1.875-1.875v-6.198a2.29 2.29 0 0 0 .091-.086L12 5.432Z" />
               </svg>
               <p>داشبورد</p>
-            </a>
+            </Link>
           </li>
           <li>
-            <a>
+            <Link to={"/CreateNewProduct"}>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 24 24"
@@ -62,10 +62,10 @@ function Admin() {
                 />
               </svg>
               <p>محصول جدید</p>
-            </a>
+            </Link>
           </li>
           <li>
-            <a>
+            <Link to={"/allusers"}>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 24 24"
@@ -80,10 +80,10 @@ function Admin() {
                 <path d="M5.082 14.254a8.287 8.287 0 0 0-1.308 5.135 9.687 9.687 0 0 1-1.764-.44l-.115-.04a.563.563 0 0 1-.373-.487l-.01-.121a3.75 3.75 0 0 1 3.57-4.047ZM20.226 19.389a8.287 8.287 0 0 0-1.308-5.135 3.75 3.75 0 0 1 3.57 4.047l-.01.121a.563.563 0 0 1-.373.486l-.115.04c-.567.2-1.156.349-1.764.441Z" />
               </svg>
               <p>مدیریت کاربران</p>
-            </a>
+            </Link>
           </li>
           <li>
-            <a>
+            <Link to={"/orders"}>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 24 24"
@@ -102,10 +102,10 @@ function Admin() {
                 />
               </svg>
               <p>سفارشات</p>
-            </a>
+            </Link>
           </li>
           <li>
-            <a>
+            <Link to={"/Profile"}>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
                 viewBox="0 0 24 24"
@@ -119,7 +119,7 @@ function Admin() {
                 />
               </svg>
               <p>پروفایل</p>
-            </a>
+            </Link>
           </li>
           <li>
             <Link to={"/"} onClick={handleLogout}>
