--- src/components/Layout/AuthPageImage.tsx
@@ -1,9 +1,0 @@
-const AuthPageImage = () => {
-  return (
-    <>
-      <img src="" alt="" />
-    </>
-  );
-};
-
-export default AuthPageImage;
