+++ src/components/Layout/Layout.tsx
@@ -1,15 +1,11 @@
-// import User from "./Layout/User";
 import Menu from "./Menu";
 import { Outlet } from "react-router";
-// import Admin from "./Layout/Admin";
 
 const Layout = () => {
   return (
     <div>
       <aside className="fixed z-10 right-0 min-h-screen p-2 bg-base-300">
         <Menu />
-        {/* <User /> */}
-        {/* <Admin /> */}
       </aside>
       <main className="min-h-screen">
         <Outlet />
