--- src/context/AuthContext.tsx
@@ -1,40 +1,0 @@
-// import React, { createContext, useContext, useState, useEffect } from "react";
-
-// interface AuthContextType {
-//   token: string | null;
-//   login: (token: string) => void;
-//   logout: () => void;
-// }
-
-// const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-// export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
-//   children,
-// }) => {
-//   const [token, setToken] = useState<string | null>(() =>
-//     localStorage.getItem("token")
-//   );
-
-//   const login = (newToken: string) => {
-//     setToken(newToken);
-//     localStorage.setItem("token", newToken);
-//   };
-
-//   const logout = () => {
-//     setToken(null);
-
-//     localStorage.removeItem("token");
-//   };
-
-//   return (
-//     <AuthContext.Provider value={{ token, login, logout }}>
-//       {children}
-//     </AuthContext.Provider>
-//   );
-// };
-
-// export const useAuth = () => {
-//   const context = useContext(AuthContext);
-//   if (!context) throw new Error("useAuth must be used within AuthProvider");
-//   return context;
-// };
