+++ src/main.tsx
@@ -7,10 +7,6 @@ import App from "./App.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    {/* <BrowserRouter>
-      <AuthProvider> */}
     <App />
-    {/* </AuthProvider>
-    </BrowserRouter> */}
   </StrictMode>
 );
